+++ src/test/java/org/apiphany/lang/retry/WaitTimeoutTest.java
@@ -2,6 +2,7 @@ package org.apiphany.lang.retry;
 
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.greaterThanOrEqualTo;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
@@ -9,7 +10,9 @@ import java.time.Duration;
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
 import java.util.concurrent.TimeUnit;
+import java.util.function.BiConsumer;
 
+import org.apiphany.lang.Holder;
 import org.junit.jupiter.api.Test;
 import org.morphix.lang.thread.Threads;
 
@@ -129,15 +132,136 @@ class WaitTimeoutTest {
 	}
 
 	@Test
+	void shouldReturnFalseOnEqualsIfOtherIsNull() {
+		WaitTimeout waitTimeout = WaitTimeout.of(TIMEOUT, INTERVAL);
+
+		boolean result = waitTimeout.equals(null);
+
+		assertFalse(result);
+	}
+
+	@Test
+	void shouldReturnFalseOnEqualsIfOtherIsADifferentClass() {
+		WaitTimeout waitTimeout = WaitTimeout.of(TIMEOUT, INTERVAL);
+
+		boolean result = waitTimeout.equals(new Object());
+
+		assertFalse(result);
+	}
+
+	@Test
 	void shouldSetTheStartTimeOnCallingStart() {
 		WaitTimeout waitTimeout = WaitTimeout.of(TIMEOUT, INTERVAL);
 
 		waitTimeout.start();
-		Threads.safeSleep(Duration.ofMillis(50));
+		Threads.safeSleep(Duration.ofMillis(5));
 
 		WaitTimeout copyWithoutStart = waitTimeout.copy();
 		boolean result = waitTimeout.equals(copyWithoutStart);
 
 		assertFalse(result);
+	}
+
+	@Test
+	void shouldSetTheProvidedStartTimeOnCallingStart() {
+		WaitTimeout waitTimeout = WaitTimeout.of(TIMEOUT, INTERVAL);
+
+		waitTimeout.start(START);
+
+		WaitTimeout copyWithStart = waitTimeout.copy();
+		copyWithStart.start(START);
+		boolean result = waitTimeout.equals(copyWithStart);
+
+		assertTrue(result);
+	}
+
+	@Test
+	void shouldReturnFalseOnKeepWaitingIfTimeoutReached() {
+		WaitTimeout waitTimeout = WaitTimeout.of(Duration.ofSeconds(-10), INTERVAL);
+		waitTimeout.start();
+
+		boolean result = waitTimeout.keepWaiting();
+
+		assertFalse(result);
+	}
+
+	@Test
+	void shouldReturnTrueOnKeepWaitingIfTimeoutNotReached() {
+		WaitTimeout waitTimeout = WaitTimeout.of(Duration.ofSeconds(10), INTERVAL);
+		waitTimeout.start();
+
+		boolean result = waitTimeout.keepWaiting();
+
+		assertTrue(result);
+	}
+
+	@Test
+	void shouldReturnTrueOnIsOverWithEpochMillisWhenTimeoutReached() {
+		WaitTimeout waitTimeout = WaitTimeout.of(Duration.ofSeconds(-10), INTERVAL);
+		waitTimeout.start();
+
+		boolean result = waitTimeout.isOver(Instant.now().toEpochMilli());
+
+		assertTrue(result);
+	}
+
+	@Test
+	void shouldReturnFalseOnIsOverWithEpochMillisWhenTimeoutNotReached() {
+		WaitTimeout waitTimeout = WaitTimeout.of(Duration.ofSeconds(10), INTERVAL);
+		waitTimeout.start();
+
+		boolean result = waitTimeout.isOver(Instant.now().toEpochMilli());
+
+		assertFalse(result);
+	}
+
+	@Test
+	void shouldReturnTrueOnIsOverWhenTimeoutReached() {
+		WaitTimeout waitTimeout = WaitTimeout.of(Duration.ofSeconds(-10), INTERVAL);
+		waitTimeout.start();
+
+		boolean result = waitTimeout.isOver(Instant.now());
+
+		assertTrue(result);
+	}
+
+	@Test
+	void shouldReturnFalseOnIsOverWhenTimeoutNotReached() {
+		WaitTimeout waitTimeout = WaitTimeout.of(Duration.ofSeconds(10), INTERVAL);
+		waitTimeout.start();
+
+		boolean result = waitTimeout.isOver(Instant.now());
+
+		assertFalse(result);
+	}
+
+	@Test
+	void shouldWaitOnNow() {
+		WaitTimeout waitTimeout = WaitTimeout.of(Duration.ofMillis(10), Duration.ofMillis(5));
+		waitTimeout.start();
+
+		Instant before = Instant.now();
+		waitTimeout.now();
+		Instant after = Instant.now();
+
+		long elapsedMillis = Duration.between(before, after).toMillis();
+		assertThat(elapsedMillis, greaterThanOrEqualTo(5L));
+	}
+
+	@Test
+	void shouldNotWaitOnNowIfTimeoutReached() {
+		Holder<Boolean> sleepCalls = Holder.of(Boolean.FALSE);
+		BiConsumer<Long, TimeUnit> sleepAction = (interval, timeUnit) -> sleepCalls.setValue(Boolean.TRUE);
+		WaitTimeout waitTimeout = new WaitTimeout(-10, TimeUnit.SECONDS, -10, TimeUnit.SECONDS) {
+			@Override
+			public BiConsumer<Long, TimeUnit> sleepAction() {
+				return sleepAction;
+			}
+		};
+		waitTimeout.start();
+
+		waitTimeout.now();
+
+		assertThat(sleepCalls.getValue(), equalTo(Boolean.FALSE));
 	}
 }